+++ index.ts
@@ -11,8 +11,14 @@ export {
 export { traceEvent } from "./debug/traceEvent"
 export { createExpressServer } from "./http/expressAdapter"
 export { createFastifyServer } from "./http/fastifyAdapter"
-export { type AdapterOptions, sanitizeErrorMessage, withTimeout } from "./http/timeout"
+export {
+  type AdapterOptions,
+  buildErrorBody,
+  sanitizeErrorMessage,
+  withTimeout,
+} from "./http/timeout"
 export { traceMiddleware } from "./middleware/traceMiddleware"
+export { type CorsOptions, corsPlugin } from "./plugins/corsPlugin"
 export { loggerPlugin } from "./plugins/loggerPlugin"
 export { type OtelPluginOptions, otelPlugin } from "./plugins/otelPlugin"
 export { type Plugin, PluginManager } from "./plugins/PluginManager"